Expense_Tracker/Backend/requirements.txt
fastapi
uvicorn
motor
passlib[bcrypt]
python-jose[cryptography]
python-multipart
email-validator
pydantic[email]
python-dotenv
dnspython
